Move outbox state into own file
