+++ packages/wallet/src/redux/outbox/sharedState.ts
@@ -1,0 +1,8 @@
+import { TransactionRequest } from 'ethers/providers';
+import { WalletEvent, DisplayAction } from 'magmo-wallet-client';
+
+export interface OutboxState {
+  displayOutbox?: DisplayAction;
+  messageOutbox?: WalletEvent;
+  transactionOutbox?: TransactionRequest;
+}
