+++ packages/wallet/src/redux/reducers/__tests__/helpers.ts
@@ -1,5 +1,6 @@
 import { ChannelState } from '../../channel/state';
-import { NextChannelState, OutboxState } from '../../sharedState';
+import { OutboxState } from '../../outbox/sharedState';
+import { NextChannelState } from '../../sharedState';
 import { Commitment } from 'fmg-core';
 import * as outgoing from 'magmo-wallet-client/lib/wallet-events';
